Revert "Make sure matplotlib is imported before openseespy"
Turns out this didn't work; the workaround only works if matplotlib is
imported first before *all* of OpenSeesTools; looks like the Python
import mechanism traverses all the paths instead of doing things simply
one at a time. The only fix I've found so far is to simply remove the
offending libraries from the openseespy folder, but that's not a easily
distributable fix.