Handle opensees versions a little better
OpenSees is now imported by the main __init__ file, and all the other
modules import it from there. I think this should make it so that you
can specify the OpenSeesPy module to use by setting that variable on the
OpenSeesTools module. And even if not, it makes things a little neater
than importing from basic.